DP-158 move installation of child (jvm worker) signal handlers after Daemon.start().
This is what the original jsvc did and it supports applications that never return from Daemon.start()...  since the signal handler is never installed the default signal handler will likely terminate the process on receipt of signal.   The HDFS Datanode is one of the apps that never returns from Daemon.start().